Update patch set 2
Patch Set 2:

Well, I will try to answer. Murano - the service which allow uploading special packages - Murano applications, which deliveded via zip archives. So, the extra dir contain a set of packages for the tests. By default (if used installation from sources), we are chown murano_tempest_tests dir to the tempest user (devstack installation).
But when murano is installed from packages, this dir appears in system python forder. Then, Alexey made a patch for murano to use temp dirs for archives and actions with them. So, I think this is ok behavior for now.

Patch-set: 2
Label: Code-Review=0