Update patch set 3
Patch Set 3: Code-Review+1

It just needed to enable testing for the installed packages by puppets, AFAIK. 
Murano is not a core OpenStack project like a glance or nova. So, there is no way to integrate it into tempest as is. We have only the tempest plugin.
So, the tempest plugin uses extra folder in the murano_tempest_tests to create zip archives(with murano packages). It's just as designed, that murano takes zip on input. So, we need to create and store these packages to be able to upload them..

Patch-set: 3
Reviewer: Gerrit User 258 <258@270e2033-b340-4cff-9539-693957ebf0e7>
Label: Code-Review=+1